Switch to a central registry for C++-owned VALUEs
In order to play nice with Ruby's garbage collection, while still
having only one C++ representation of a single ruby object, it's
necessary to keep track of the association between Ruby objects and
their C++ wrappers. This is done with a new `value_registry` type,
which `value_wrapper` is intimately aware of.

Constructing C++ wrappers for ruby objects is now done through a
`value_wrapper::wrap` convenience function, which talks to the
value_registry to figure out what's up.